constant-time 모듈 분할과 트레이트 개명 - internal 을 x86_64 aarch64 fallback ct128 로 분할 - 트레이트를 traits 모듈로 분리하고 CtGreeter 를 CtGtOps 로 개명 - 메서드명을 ct_eq ct_ne ct_gt ct_lt 로 통일해 표준 트레이트와 충돌 제거 - 배열과 Secret 에 대한 CtEqOps 구현 추가 - swap 임시값 소거를 zeroize 크레이트로 위임 - check_ct_asm 에 배열 ct_eq 분기 상한과 memcmp 치환 검사 추가 - 전 크레이트 호출부와 README 갱신
