+++ constant-time/Cargo.toml
@@ -6,10 +6,10 @@ authors.workspace = true
 license.workspace = true
 
 [dependencies]
+zeroize = { workspace = true }
 
 [dev-dependencies]
 rng = { workspace = true }
-zeroize = { workspace = true }
 # Phase 4 Plan 02 wire host 단위테스트 9 종 추가 (postcard bitexact + Blake3 cross-validation)
 blake = { workspace = true }
 postcard = { version = "1", default-features = false }
