+++ constant-time/README.md
@@ -12,9 +12,9 @@
 
 - `Choice`: 항상 0 또는 1 값을 갖는 상수-시간 bool. `&`, `|`, `^`, `!` 비트 연산이 정규화 없이 0/1 불변을 보존하며, 민감 값 유출(CWE-532) 방지를 위해 `Debug`를 의도적으로 파생하지 않습니다.
 - `CtSelOps`: 조건 선택(`select`), 조건 대입(`assign`), 조건 교환(`swap`).
-- `CtEqOps`: 동등 비교(`eq`, `ne`).
-- `CtGreeter`: 대소 비교(`gt`).
-- `CtLess`: 작음 비교(`lt`). `CtEqOps + CtGreeter`를 만족하는 모든 타입에 `!gt & !eq`로 자동 제공됩니다.
+- `CtEqOps`: 동등 비교(`ct_eq`, `ct_ne`).
+- `CtGtOps`: 대소 비교(`ct_gt`).
+- `CtLess`: 작음 비교(`ct_lt`). `CtEqOps + CtGtOps`를 만족하는 모든 타입에 `!ct_gt & !ct_eq`로 자동 제공됩니다.
 
 네 가지 트레이트는 고정폭 정수 `u8`부터 `i128`까지에 구현되며, `Sealed`로 봉인되어 외부 타입이 끼어들 수 없습니다. 각 구현은 `internal` 모듈의 저수준 프리미티브에 위임합니다.
 
@@ -47,7 +47,7 @@ x86_64와 aarch64에서 각 프리미티브는 인-라인 어셈블리로 직접
 
 ### 3. Choice 0/1 불변
 
-`Choice`는 `(v | v.wrapping_neg()) >> 7` 마스크로 분기 없이 0/1로 정규화되고, 비트 연산이 이 불변을 보존합니다. 덕분에 `lt = !gt & !eq` 같은 상위 조합도 분기 없이 유지됩니다.
+`Choice`는 `(v | v.wrapping_neg()) >> 7` 마스크로 분기 없이 0/1로 정규화되고, 비트 연산이 이 불변을 보존합니다. 덕분에 `lt = !ct_gt & !ct_eq` 같은 상위 조합도 분기 없이 유지됩니다.
 
 추가로, 검증된 인-라인 어셈블리가 없는 아키텍처는 컴파일 게이트(`compile_error!`)로 거부되어 best-effort `black_box` fallback이 고보안 빌드에 섞이지 않습니다(아래 '발견한 문제와 조치' 절 참고). `swap`의 영점화 루프는 종료 조건이 `size_of::<Self>()`라는 컴파일타임 상수이므로 그 분기는 비밀에 의존하지 않습니다.
