+++ sha2/src/sha2_256.rs
@@ -1,5 +1,6 @@
 use crate::{Digest, SHA256State};
-use constant_time::{Choice, CtEqOps, CtGreeter, CtSelOps};
+use constant_time::Choice;
+use constant_time::traits::{CtEqOps, CtGtOps, CtSelOps};
 use zeroize::{Secret, Zeroize};
 
 const SHA_256_K: [u32; 64] = [
@@ -132,7 +133,7 @@ impl SHA256State {
             //   select(a, b, choice) -> choice==1일 때 b, choice==0일 때 a
             //   is_ge==1 (remain >= fill) -> fill   (현재 블록 완료)
             //   is_ge==0 (remain <  fill) -> remain (남은 것을 가져옴)
-            let is_ge: Choice = CtGreeter::gt(&remain, &fill) | CtEqOps::eq(&remain, &fill);
+            let is_ge: Choice = CtGtOps::ct_gt(&remain, &fill) | CtEqOps::ct_eq(&remain, &fill);
             let chunk_len: usize = usize::select(&remain, &fill, is_ge);
 
             self.buffer[self.buffer_len..self.buffer_len + chunk_len]
@@ -164,7 +165,7 @@ impl SHA256State {
         // 이 블록(위치 56-63)에서
         //   needs_extra == 1  -> buffer_len > 56: 길이는 두 번째 블록으로 가야 함
         //   needs_extra == 0  -> buffer_len ≤ 56: 길이는 이 블록에 맞음
-        let needs_extra: Choice = CtGreeter::gt(&self.buffer_len, &56usize);
+        let needs_extra: Choice = CtGtOps::ct_gt(&self.buffer_len, &56usize);
         let not_extra: Choice = !needs_extra;
 
         let total_len_bytes = self.total_len.to_be_bytes();
